add option keep_xtc in gromacs engine
When studying large systems we might want to keep
the compressed coordinates in the .xtc files. Using
keep_old = true in the .toml now also puts these
trajectories into the accepted/ folder. They will not
be deleted even when delete_old = true. This should
also work with compressed-x-grps to select a subset of
atoms, if one is interested in higher timestep resolutions
than subcycles